drivers:platform:aducm3029: Rename uart_stdio driver
Rename uart_stdio driver to aducm3029_uart_stdio and adapt
projects accordingly.
