+++ drivers/platform/aducm3029/aducm3029_uart_stdio.h
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   uart_stdio.c
- *   @brief  Source file to use printf over uart
+ *   @file   aducm3029_uart_stdio.h
+ *   @brief  Header file for UART driver stdout/stdin redirection.
  *   @author Mihail Chindris (mihail.chindris@analog.com)
 ********************************************************************************
  * Copyright 2021(c) Analog Devices, Inc.
@@ -36,98 +36,11 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
+#ifndef _ADUCM3029_UART_STDIO_H_
+#define _ADUCM3029_UART_STDIO_H_
 
-#if ENABLE_UART_STDIO
-#include <stdio.h>
-#include <errno.h>
-#include <sys/stat.h>
 #include "no_os_uart.h"
 
-static struct no_os_uart_desc *g_uart;
+void init_uart_stdio(struct no_os_uart_desc *desc);
 
-#define STDIN_FILENO  0
-#define STDOUT_FILENO 1
-#define STDERR_FILENO 2
-
-void init_uart_stdio(struct no_os_uart_desc *desc)
-{
-	g_uart = desc;
-}
-
-int _isatty(int fd)
-{
-	if (fd >= STDIN_FILENO && fd <= STDERR_FILENO)
-		return 1;
-
-	errno = EBADF;
-	return 0;
-}
-
-int _write(int fd, char* ptr, int len)
-{
-	int32_t ret;
-
-	if (!g_uart)
-		return EIO;
-
-	if (fd == STDOUT_FILENO || fd == STDERR_FILENO) {
-		ret = no_os_uart_write(g_uart, (uint8_t *) ptr, len);
-		if (ret < 0) {
-			errno = ret;
-			return EIO;
-		}
-		return len;
-	}
-	errno = EBADF;
-
-	return -1;
-}
-
-int _close(int fd)
-{
-	if (fd >= STDIN_FILENO && fd <= STDERR_FILENO)
-		return 0;
-
-	errno = EBADF;
-	return -1;
-}
-
-int _lseek(int fd, int ptr, int dir)
-{
-	(void) fd;
-	(void) ptr;
-	(void) dir;
-
-	errno = EBADF;
-	return -1;
-}
-
-int _read(int fd, char* ptr, int len)
-{
-	int32_t ret;
-
-	if (!g_uart)
-		return EIO;
-
-	if (fd == STDIN_FILENO) {
-		ret = no_os_uart_read(g_uart, (uint8_t *) ptr, len);
-		if (ret < 0)
-			return EIO;
-
-		return len;
-	}
-	errno = EBADF;
-	return -1;
-}
-
-int _fstat(int fd, struct stat* st)
-{
-	if (fd >= STDIN_FILENO && fd <= STDERR_FILENO) {
-		st->st_mode = S_IFCHR;
-		return 0;
-	}
-
-	errno = EBADF;
-	return -1;
-}
-#endif
+#endif //_ADUCM3029_UART_STDIO_H_
