+++ projects/cn0565/src.mk
@@ -59,11 +59,11 @@ SRCS += $(PLATFORM_DRIVERS)/delay.c \
 	$(PLATFORM_DRIVERS)/aducm3029_gpio_irq.c \
 	$(PLATFORM_DRIVERS)/aducm3029_rtc.c \
 	$(PLATFORM_DRIVERS)/aducm3029_uart.c \
-	$(PLATFORM_DRIVERS)/uart_stdio.c \
+	$(PLATFORM_DRIVERS)/aducm3029_uart_stdio.c \
 	$(PLATFORM_DRIVERS)/platform_init.c
 
 INCS +=	$(INCLUDE)/no_os_delay.h \
-	$(PLATFORM_DRIVERS)/uart_stdio.h \
+	$(PLATFORM_DRIVERS)/aducm3029_uart_stdio.h \
 	$(PLATFORM_DRIVERS)/uart_extra.h \
 	$(PLATFORM_DRIVERS)/aducm3029_irq.h \
 	$(PLATFORM_DRIVERS)/aducm3029_gpio_irq.h \
